backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
chromadb==0.4.18
pypdf2==3.0.1
python-dotenv==1.0.0
pydantic==2.5.0
langchain==0.0.350
sentence-transformers>=2.5.0
huggingface-hub>=0.19.0,<1.0.0
numpy==1.24.3
pandas==2.0.3
groq==0.31.0
httpx>=0.24.0
urllib3>=2.0.0,<3.0.0
requests>=2.25.0
